backend/requirements.txt
fastapi
uvicorn[standard]
neo4j
requests
pydantic
python-dotenv
scikit-learn
neo4j-gds
graphdatascience
torch
torch-geometric
